# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vullify API
  description: |
    REST API for container registry integration, vulnerability scans (Trivy), findings,
    enrichment (EPSS / KEV), and reporting. List responses wrap payloads in `data` with optional `meta` pagination.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/registries:
    get:
      tags: [Registries]
      summary: List registries
      operationId: listRegistries
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryListEnvelope"
    post:
      tags: [Registries]
      summary: Create registry
      description: Validates connector credentials with a live connection test before insert.
      operationId: createRegistry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          description: Registry connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/registries/{id}:
    parameters:
      - $ref: "#/components/parameters/RegistryId"
    get:
      tags: [Registries]
      summary: Get registry
      operationId: getRegistry
      responses:
        "200":
          description: Registry detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Registries]
      summary: Update registry
      operationId: updateRegistry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Registry connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    delete:
      tags: [Registries]
      summary: Soft-delete registry
      operationId: deleteRegistry
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/registries/{id}/sync:
    parameters:
      - $ref: "#/components/parameters/RegistryId"
    post:
      tags: [Registries]
      summary: Sync images from registry
      description: Lists repositories and tags from the registry and upserts image rows.
      operationId: syncRegistry
      responses:
        "200":
          description: Sync summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      synced_images:
                        type: integer
                      repositories_seen:
                        type: integer
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Registry sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/images:
    get:
      tags: [Images]
      summary: List images
      operationId: listImages
      parameters:
        - name: registry_id
          in: query
          description: Filter by registry UUID
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/images/{id}:
    parameters:
      - $ref: "#/components/parameters/ImageId"
    get:
      tags: [Images]
      summary: Get image detail
      operationId: getImage
      responses:
        "200":
          description: Image with registry URL and latest scan hints
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ImageDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans:
    post:
      tags: [Scans]
      summary: Create scan
      description: Creates a scan job and enqueues it on Redis for the worker.
      operationId: createScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScan"
      responses:
        "201":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetailEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/scans/{id}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Get scan detail
      operationId: getScan
      responses:
        "200":
          description: Scan with severity counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetailEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{id}/findings:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: List findings for scan
      operationId: listScanFindings
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: package_name
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/scans/{id}/sbom:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Get raw SBOM JSON
      description: Returns JSON body; `X-SBOM-Format` header indicates format.
      operationId: getScanSBOM
      responses:
        "200":
          description: SBOM payload
          headers:
            X-SBOM-Format:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/findings/{id}:
    parameters:
      - $ref: "#/components/parameters/FindingId"
    get:
      tags: [Findings]
      summary: Get finding detail
      description: Includes enrichment fields when available (EPSS, KEV, known exploits).
      operationId: getFinding
      responses:
        "200":
          description: Finding detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingDetailEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/reports/vulnerability:
    get:
      tags: [Reports]
      summary: Vulnerability occurrence report
      operationId: vulnerabilityReport
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: from
          in: query
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date (YYYY-MM-DD), inclusive end-of-day
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated vulnerability rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityReportEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/scans/adhoc:
    post:
      tags: [Scans]
      summary: Create ad-hoc scan
      description: |
        Accepts an image reference directly (no pre-existing image or registry required).
        Upserts the image record, creates a scan, and enqueues it on Redis for the worker.
      operationId: createAdHocScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdHocScanRequest"
      responses:
        "201":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageScanReportEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/suppressions:
    get:
      tags: [Suppressions]
      summary: List suppressions
      operationId: listSuppressions
      parameters:
        - name: cve_id
          in: query
          schema:
            type: string
        - name: target_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated suppressions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppressionListEnvelope"
    post:
      tags: [Suppressions]
      summary: Create suppression
      description: Suppresses a specific CVE for a package/target combination.
      operationId: createSuppression
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuppressionCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuppressionEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/suppressions/{id}:
    parameters:
      - $ref: "#/components/parameters/SuppressionId"
    delete:
      tags: [Suppressions]
      summary: Revoke suppression
      operationId: deleteSuppression
      responses:
        "204":
          description: Revoked
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/dashboard/cves:
    get:
      tags: [Dashboard]
      summary: Global CVE view
      description: |
        Returns unique CVEs across all scans, deduplicated, with aggregated risk scores.
        Supports filtering by severity and sorting by risk_score.
      operationId: dashboardCVEs
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort field (risk_score, severity, last_seen)
          schema:
            type: string
            default: risk_score
        - name: sort_order
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated global CVE results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalCVEResultEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    RegistryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ImageId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ScanId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FindingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SuppressionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error:
              code: VALIDATION_ERROR
              message: invalid JSON body
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    Meta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          format: int64
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/Meta"
    Registry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          description: dockerhub, gitlab, ecr, gcr, ghcr
        url:
          type: string
        credentials_set:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RegistryCreate:
      type: object
      required: [name, type, url]
      properties:
        name:
          type: string
        type:
          type: string
        url:
          type: string
        credentials:
          description: Connector-specific JSON (default {})
          type: object
    RegistryUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        url:
          type: string
        credentials:
          type: object
    RegistryEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Registry"
    RegistryListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Registry"
    Image:
      type: object
      properties:
        id:
          type: string
          format: uuid
        registry_id:
          type: string
          format: uuid
        repository:
          type: string
        tag:
          type: string
        digest:
          type: string
          nullable: true
        first_seen_at:
          type: string
        last_seen_at:
          type: string
    ImageDetail:
      allOf:
        - $ref: "#/components/schemas/Image"
        - type: object
          properties:
            registry_url:
              type: string
            latest_scan_id:
              type: string
              nullable: true
            latest_scan_status:
              type: string
              nullable: true
            latest_scan_started_at:
              type: string
              nullable: true
            latest_scan_completed_at:
              type: string
              nullable: true
    ImageListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Image"
    CreateScan:
      type: object
      description: Provide exactly one of image_id or image_ref
      properties:
        image_id:
          type: string
          format: uuid
        image_ref:
          type: string
          description: Canonical pull reference matching stored images
    ScanDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        status:
          type: string
        triggered_by:
          type: string
        started_at:
          type: string
          nullable: true
        completed_at:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        trivy_version:
          type: string
          nullable: true
        repository:
          type: string
        tag:
          type: string
        severity_counts:
          type: object
          additionalProperties:
            type: integer
            format: int64
    ScanDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ScanDetail"
    FindingListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vulnerability_id:
          type: string
        package_name:
          type: string
        severity:
          type: string
        created_at:
          type: string
        installed_version:
          type: string
        fixed_version:
          type: string
        title:
          type: string
    FindingListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/FindingListItem"
    FindingDetail:
      type: object
      description: Dynamic object; optional enrichment fields
      additionalProperties: true
    FindingDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/FindingDetail"
    VulnerabilityReportRow:
      type: object
      properties:
        vulnerability_id:
          type: string
        severity:
          type: string
        occurrences:
          type: integer
        last_seen:
          type: string
    VulnerabilityReportEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/VulnerabilityReportRow"
...
    Suppression:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cve_id:
          type: string
        pkg_name:
          type: string
        target_id:
          type: string
          format: uuid
          nullable: true
          description: NULL means global suppression
        reason:
          type: string
        accepted_by:
          type: string
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    SuppressionCreate:
      type: object
      required: [cve_id, reason]
      properties:
        cve_id:
          type: string
        pkg_name:
          type: string
        target_id:
          type: string
          format: uuid
        reason:
          type: string
        expires_at:
          type: string
          format: date-time
    SuppressionEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Suppression"
    SuppressionListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Suppression"
    AdHocScanRequest:
      type: object
      required: [image_ref]
      properties:
        image_ref:
          type: string
          description: Fully qualified image reference (e.g. alpine:3.19, ghcr.io/org/repo:tag)
        registry_id:
          type: string
          format: uuid
          description: Optional registry to associate with the image
    ImageScanReport:
      type: object
      properties:
        scan_id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        status:
          type: string
        image_ref:
          type: string
        repository:
          type: string
        tag:
          type: string
    ImageScanReportEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ImageScanReport"
    GlobalCVEResult:
      type: object
      properties:
        vulnerability_id:
          type: string
          description: CVE identifier
        severity:
          type: string
        title:
          type: string
        affected_images:
          type: integer
        last_seen:
          type: string
        epss_score:
          type: number
          nullable: true
        epss_percentile:
          type: number
          nullable: true
        kev_listed:
          type: boolean
        risk_score:
          type: number
          nullable: true
    GlobalCVEResultEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/GlobalCVEResult"